+++ app/src/main/AndroidManifest.xml
@@ -34,11 +34,11 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".CountryNavigationActivity"
+            android:name=".preference.PreferenceNavigationActivity"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustPan" />
         <activity
-            android:name=".support.CreditActivity"
+            android:name="com.delacrixmorgan.squark.preference.credit.CreditActivity"
             android:screenOrientation="portrait" />
     </application>
 </manifest>
