+++ app/src/main/java/com/delacrixmorgan/squark/preference/SupportListFragment.kt
@@ -1,4 +1,4 @@
-package com.delacrixmorgan.squark.support
+package com.delacrixmorgan.squark.preference
 
 import android.os.Bundle
 import android.view.LayoutInflater
