+++ app/src/main/java/com/delacrixmorgan/squark/preference/country/CountryListFragment.kt
@@ -1,4 +1,4 @@
-package com.delacrixmorgan.squark.country
+package com.delacrixmorgan.squark.preference.country
 
 import android.app.Activity
 import android.content.SharedPreferences
