+++ app/src/main/java/com/delacrixmorgan/squark/preference/country/CountryListListener.kt
@@ -1,4 +1,4 @@
-package com.delacrixmorgan.squark.country
+package com.delacrixmorgan.squark.preference.country
 
 import com.delacrixmorgan.squark.data.model.Country
